SWModule: Removed broken terminateSearch functionality.
It was totally broken. First of all, the boolean field was not atomic, hence the
effect of changing its value might not have reached the other thread. If it were
made atomic, the effect would reach the other thread, but there is a chance that
the effect would get ignored due to terminateSearch being reset to false first.
Thirdly, any third thread calling search() will also race to annul the effect.