SWModule: Removed the operators for readability.
I'm not sure the semantics for such operators are straightforward to understand.
In either case, removing these seems to simplify things. In the future, we might
have to rename the related virtual methods anyway.

Overall this is more of a "remove these until I better understand the code"-type
of change.