SWText: Simplified handling of tmpVK1 and tmpVK2 fields.
Calling createKey() in the constructor acts as if staticCreateKey was called
because during SWText construction the vtable points to SWText::createKey()
anyway.

Secondly, after this change the mutable keyword for the tmpVK1 and tmpVK2 fields
actually does make some sense. It previously only made the pointers mutable, and
not the objects they pointed to. However, nothing tried to mutate the pointers.