keys: Removed namespace Position and related assignment operators on keys.
These really don't make much sense as assignment operators, because this kind of
assignment is partial. These are more like setters. Therefore it seems better to
use the underlying methods directly, because it gives a better impression about
what actually happens: that the entire object is NOT substituted, just a part of
it is changed. Additionally, the return value was never used.