SWConfig: Fixed whitespace handling and added missing continuation handling
Handling of configuration files in this class has diverted from what it used to
be in the Sword library, including the handling of whitespace and continuation
lines.

The confusion was caused by the fact, that previously SWConfig used
FileMgr::getLine() to read lines from files. What was not obvious from this
interface was that FileMgr::getLine() also does a number of extra things, such
as stripping of whitespace, handling of continuation lines etc.

This commit attempts to fix the handling of whitespace to again match the logic
of the Sword library, as well as re-introduces the support for continuation
lines.